fix(partnercenter): bugbot
