+++ src/lib/types/index.ts
@@ -10,3 +10,4 @@ export * from './licenses.types'
 export * from './users.types'
 export * from './gdap.types'
 export * from './domains.types'
+export * from './pricing.types'
