+++ src/lib/types/licenses.types.ts
@@ -54,51 +54,3 @@ export interface UserLicenseAssignment {
     productSku: ProductSku
     attributes: Attributes
 }
-
-export interface OfferMatrixEntry {
-    ProductTitle: string
-    ProductId: string
-    SkuId: string
-    SkuTitle: string
-    ProvisioningId: string
-    ProvisioningString: string
-    MinLicenses: string
-    MaxLicenses: string
-    AssetOwnershipLimit: string
-    AssetOwnershipLimitType: string
-    ProductSkuPreRequisites: string
-    ProductSkuConversion: string
-    Description: string
-    AllowedCountries: string
-}
-
-export interface PriceSheetEntry {
-    ProductTitle: string
-    ProductId: string
-    SkuId: string
-    SkuTitle: string
-    Publisher: string
-    SkuDescription: string
-    UnitOfMeasure: string
-    TermDuration: 'P1M' | 'P1Y'
-    BillingPlan: 'Annual' | 'Monthly'
-    Market: string
-    Currency: string
-    UnitPrice: string
-    PricingTierRangeMin?: string
-    PricingTierRangeMax?: string
-    EffectiveStartDate: string
-    EffectiveEndDate: string
-    Tags: string
-    'ERP Price': string
-    Segment: 'Commercial' | 'Government' | 'Education' | 'Charity'
-}
-
-export interface PlanIdentifierEntry {
-    Product_Display_Name: string
-    String_Id: string
-    GUID: string
-    Service_Plan_Name: string
-    Service_Plan_Id: string
-    Service_Plans_Included_Friendly_Names: string
-}
